+++ src/geophires_x/EconomicsS_DAC_GT.py
@@ -366,9 +366,7 @@ class EconomicsS_DAC_GT(Economics.Economics):
                 key = ParameterToModify.Name.strip()
                 if key in model.InputParameters:
                     ParameterReadIn = model.InputParameters[key]
-                    # Before we change the paremater, let's assume that the unit preferences will match -
-                    # if they don't, the later code will fix this.
-                    ParameterToModify.CurrentUnits = ParameterToModify.PreferredUnits
+
                     # this should handle all the non-special cases
                     ReadParameter(ParameterReadIn, ParameterToModify, model)
 
@@ -678,4 +676,6 @@ class EconomicsS_DAC_GT(Economics.Economics):
         #                                              (self.CarbonExtractedAnnually.value[i] * model.economics.CarbonPrice.value[i]))
 #            if i > 0:
 #                model.economics.CarbonCummCashFlow.value[i] = model.economics.CarbonCummCashFlow.value[i - 1] + model.economics.CarbonRevenue.value[i]
-        model.logger.info("Complete " + str(__class__) + ": " + sys._getframe().f_code.co_name)
+
+        self._calculate_derived_outputs(model)
+        model.logger.info(f'Complete {str(__class__)}: {sys._getframe().f_code.co_name}')