+++ src/geophires_x/SurfacePlantDoubleFlash.py
@@ -118,7 +118,7 @@ class SurfacePlantDoubleFlash(SurfacePlant):
                                                 model.wellbores.Tinj.value, ReinjTemp, self.T_chp_bottom.value,
                                                  self.enduse_efficiency_factor.value, self.chp_fraction.value)
 
-        # subtract pumping power for net electricity and  calculate first law efficiency
+        # subtract pumping power for net electricity and calculate first law efficiency
         self.NetElectricityProduced.value = self.ElectricityProduced.value - model.wellbores.PumpingPower.value
         self.FirstLawEfficiency.value = self.NetElectricityProduced.value/HeatExtractedTowardsElectricity
 
@@ -133,4 +133,5 @@ class SurfacePlantDoubleFlash(SurfacePlant):
         self.RemainingReservoirHeatContent.value = SurfacePlant.remaining_reservoir_heat_content(
             self, model.reserv.InitialReservoirHeatContent.value, self.HeatkWhExtracted.value)
 
+        self._calculate_derived_outputs(model)
         model.logger.info(f'complete {self.__class__.__name__}: {__name__}')